Move blanks to separate package
