+++ openprocurement/auctions/dgf/tests/blanks/auction_blanks.py
