+++ openprocurement/auctions/dgf/tests/blanks/bidder_blanks.py
