+++ openprocurement/auctions/dgf/tests/blanks/contract_blanks.py
