+++ openprocurement/auctions/dgf/tests/blanks/document_blanks.py
