+++ openprocurement/auctions/dgf/tests/blanks/lot_blanks.py
