+++ openprocurement/auctions/dgf/tests/blanks/tender_blanks.py
