+++ openprocurement/auctions/dgf/tests/complaint.py
@@ -7,7 +7,7 @@ from openprocurement.auctions.dgf.tests.base import (
     BaseAuctionWebTest, test_lots,
     test_financial_auction_data, test_financial_organization
 )
-from openprocurement.auctions.dgf.tests.complaint_blanks import (
+from openprocurement.auctions.dgf.tests.blanks.complaint_blanks import (
     # AuctionComplaintResourceTest
     create_auction_complaint_invalid,
     create_auction_complaint,
