+++ openprocurement/auctions/dgf/tests/document.py
@@ -6,7 +6,7 @@ from openprocurement.auctions.core.tests.base import snitch
 from openprocurement.auctions.dgf.tests.base import (
     BaseAuctionWebTest,  test_financial_auction_data
 )
-from openprocurement.auctions.dgf.tests.document_blanks import (
+from openprocurement.auctions.dgf.tests.blanks.document_blanks import (
     # AuctionDocumentResourceTest
     not_found,
     create_auction_document,
